Separate pathname pattern mode logic and watch mode
